requirements.txt
anthropic==0.19.0
flask==2.3.3
flask-cors==4.0.0
python-dotenv==1.0.0
requests==2.31.0
gunicorn==21.2.0